backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy==1.26.3
shapely==2.0.2
scikit-image==0.22.0
scipy==1.11.4
python-multipart==0.0.6
rasterio>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
# Real satellite data integration
pystac-client==0.7.6
planetary-computer==1.0.0
requests==2.31.0
# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
